+++ src/parsing.jl
@@ -292,7 +292,7 @@ function parse_autovec(tidy_expr::Union{Expr,Symbol})
     elseif @capture(x, fn_(args__))
 
       # `in` should be vectorized so do not add to this exclusion list
-      if fn in [:Ref :Set :Cols :(:) :∘ :repeat :across :desc :mean :std :var :median :first :last :minimum :maximum :sum :length :skipmissing :quantile :passmissing]
+      if fn in [:Ref :Set :Cols :(:) :∘ :ntile :repeat :across :desc :mean :std :var :median :first :last :minimum :maximum :sum :length :skipmissing :quantile :passmissing]
         return x
       elseif contains(string(fn), r"[^\W0-9]\w*$") # valid variable name
         return :($fn.($(args...)))